requirements.txt
# Inference & CV
ultralytics>=8.4.48
onnxruntime>=1.18.0
numpy>=1.26.0,<2.0.0
shapely>=2.0.4

# API & async
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
websockets>=12.0

# Storage
aiosqlite>=0.20.0

# Config / utils
pyyaml>=6.0.1
python-multipart>=0.0.9
